// web/src/api/generated/ExpiringApp.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

/**
 * An app close to its expiry.
 */
export type ExpiringApp = { name: string, display_name: string, expire_height: number, blocks_left: number, 
/**
 * Estimated time of the expiry block.
 */
expire_ms: number, instances: number, };